Removed BaseLeftTuple altogether, compiling but tests aren't passing * Use invokeVirtual on BaseTuple while generating ASM
