+++ drools-core/src/main/java/org/drools/core/common/InternalFactHandle.java
@@ -22,7 +22,6 @@ import java.util.function.Predicate;
 
 import org.drools.core.WorkingMemoryEntryPoint;
 import org.drools.core.factmodel.traits.TraitTypeEnum;
-import org.drools.core.reteoo.BaseLeftTuple;
 import org.drools.core.reteoo.LeftTuple;
 import org.drools.core.reteoo.RightTuple;
 import org.drools.core.rule.EntryPointId;
@@ -122,10 +121,10 @@ public interface InternalFactHandle
     boolean isPendingRemoveFromStore();
 
     void forEachRightTuple(Consumer<RightTuple> rightTupleConsumer );
-    void forEachLeftTuple(Consumer<BaseLeftTuple> leftTupleConsumer);
+    void forEachLeftTuple(Consumer<LeftTuple> leftTupleConsumer);
 
     RightTuple findFirstRightTuple(Predicate<RightTuple> rightTuplePredicate );
-    BaseLeftTuple findFirstLeftTuple(Predicate<BaseLeftTuple> lefttTuplePredicate );
+    LeftTuple findFirstLeftTuple(Predicate<LeftTuple> lefttTuplePredicate );
 
     void setFirstLeftTuple( LeftTuple firstLeftTuple );
 
@@ -156,10 +155,10 @@ public interface InternalFactHandle
         void forEachRightTuple(Consumer<RightTuple> rightTupleConsumer);
         RightTuple findFirstRightTuple(Predicate<RightTuple> rightTuplePredicate );
 
-        void forEachLeftTuple(Consumer<BaseLeftTuple> leftTupleConsumer);
-        BaseLeftTuple findFirstLeftTuple(Predicate<BaseLeftTuple> leftTuplePredicate );
+        void forEachLeftTuple(Consumer<LeftTuple> leftTupleConsumer);
+        LeftTuple findFirstLeftTuple(Predicate<LeftTuple> leftTuplePredicate );
 
-        BaseLeftTuple getFirstLeftTuple( int partition);
+        LeftTuple getFirstLeftTuple(int partition);
         void setFirstLeftTuple( LeftTuple firstLeftTuple, int partition );
 
         default LeftTuple getFirstLeftTuple(RuleBasePartitionId partitionId) {
@@ -398,7 +397,7 @@ public interface InternalFactHandle
         }
 
         @Override
-        public void forEachLeftTuple( Consumer<BaseLeftTuple> leftTupleConsumer ) {
+        public void forEachLeftTuple( Consumer<LeftTuple> leftTupleConsumer ) {
             throw new UnsupportedOperationException();
         }
 
@@ -408,7 +407,7 @@ public interface InternalFactHandle
         }
 
         @Override
-        public BaseLeftTuple findFirstLeftTuple( Predicate<BaseLeftTuple> lefttTuplePredicate ) {
+        public LeftTuple findFirstLeftTuple(Predicate<LeftTuple> lefttTuplePredicate ) {
             throw new UnsupportedOperationException();
         }
