+++ drools-core/src/main/java/org/drools/core/impl/StatefulKnowledgeSessionImpl.java
@@ -968,7 +968,7 @@ public class StatefulKnowledgeSessionImpl extends AbstractRuntime
 
         @Override
         public void execute( InternalWorkingMemory wm ) {
-            LeftInputAdapterNode lian = factHandle.getFirstLeftTuple().getTupleSource();
+            LeftInputAdapterNode lian = (LeftInputAdapterNode) factHandle.getFirstLeftTuple().getTupleSource();
             LeftInputAdapterNode.LiaNodeMemory lmem = getNodeMemory(lian);
             SegmentMemory lsmem = lmem.getSegmentMemory();
