+++ drools-core/src/main/java/org/drools/core/reteoo/ConditionalBranchNode.java
@@ -123,7 +123,7 @@ public class ConditionalBranchNode extends LeftTupleSource implements LeftTupleS
     @Override
     public LeftTuple createPeer(LeftTuple original) {
         EvalNodeLeftTuple peer = new EvalNodeLeftTuple();
-        peer.initPeer( (BaseLeftTuple) original, this );
+        peer.initPeer(original, this );
         original.setPeer( peer );
         return peer;
     }
