+++ drools-core/src/main/java/org/drools/core/reteoo/EvalConditionNode.java
@@ -146,7 +146,7 @@ public class EvalConditionNode extends LeftTupleSource
     @Override
     public LeftTuple createPeer(LeftTuple original) {
         EvalNodeLeftTuple peer = new EvalNodeLeftTuple();
-        peer.initPeer( (BaseLeftTuple) original, this );
+        peer.initPeer((LeftTuple) original, this );
         original.setPeer( peer );
         return peer;
     }
