+++ drools-core/src/main/java/org/drools/core/reteoo/ExistsNode.java
@@ -101,7 +101,7 @@ public class ExistsNode extends BetaNode {
     
     public LeftTuple createPeer(LeftTuple original) {
         NotNodeLeftTuple peer = new NotNodeLeftTuple();
-        peer.initPeer( (BaseLeftTuple) original, this );
+        peer.initPeer((LeftTuple) original, this );
         original.setPeer( peer );
         return peer;
     }
