+++ drools-core/src/main/java/org/drools/core/reteoo/JoinNodeLeftTuple.java
@@ -19,7 +19,7 @@ package org.drools.core.reteoo;
 import org.drools.core.common.InternalFactHandle;
 import org.drools.core.spi.PropagationContext;
 
-public class JoinNodeLeftTuple extends BaseLeftTuple {
+public class JoinNodeLeftTuple extends LeftTuple {
 
     private static final long serialVersionUID = 540l;
