+++ drools-core/src/main/java/org/drools/core/reteoo/NotNode.java
@@ -90,7 +90,7 @@ public class NotNode extends BetaNode {
     
     public LeftTuple createPeer(LeftTuple original) {
         NotNodeLeftTuple peer = new NotNodeLeftTuple();
-        peer.initPeer( (BaseLeftTuple) original, this );
+        peer.initPeer((LeftTuple) original, this );
         original.setPeer( peer );
         return peer;
     }
