+++ drools-core/src/main/java/org/drools/core/reteoo/QueryElementNodeLeftTuple.java
@@ -19,7 +19,7 @@ package org.drools.core.reteoo;
 import org.drools.core.common.InternalFactHandle;
 import org.drools.core.spi.PropagationContext;
 
-public class QueryElementNodeLeftTuple extends BaseLeftTuple {
+public class QueryElementNodeLeftTuple extends LeftTuple {
     private static final long serialVersionUID = 540l;
 
     public QueryElementNodeLeftTuple() {
