+++ drools-core/src/main/java/org/drools/core/reteoo/RightInputAdapterNode.java
@@ -166,7 +166,7 @@ public class RightInputAdapterNode extends ObjectSource
     
     public SubnetworkTuple createPeer(LeftTuple original) {
         SubnetworkTuple peer = new SubnetworkTuple();
-        peer.initPeer( (BaseLeftTuple) original, this );
+        peer.initPeer((LeftTuple) original, this );
         original.setPeer( peer );
         return peer;
     }
