+++ drools-core/src/main/java/org/drools/core/reteoo/RuleTerminalNodeLeftTuple.java
@@ -38,7 +38,7 @@ import org.drools.core.spi.PropagationContext;
 import org.drools.core.util.LinkedList;
 import org.kie.api.runtime.rule.FactHandle;
 
-public class RuleTerminalNodeLeftTuple<T extends ModedAssertion<T>> extends BaseLeftTuple implements
+public class RuleTerminalNodeLeftTuple<T extends ModedAssertion<T>> extends LeftTuple implements
                                                                     AgendaItem<T> {
     private static final long serialVersionUID = 540l;
     /**
