+++ drools-core/src/main/java/org/drools/core/reteoo/TimerNode.java
@@ -174,7 +174,7 @@ public class TimerNode extends LeftTupleSource
     @Override
     public LeftTuple createPeer(LeftTuple original) {
         EvalNodeLeftTuple peer = new EvalNodeLeftTuple();
-        peer.initPeer((BaseLeftTuple) original, this);
+        peer.initPeer( original, this);
         original.setPeer(peer);
         return peer;
     }
