+++ drools-mvel/src/main/java/org/drools/mvel/asm/EvalGenerator.java
@@ -92,7 +92,7 @@ public class EvalGenerator {
                     mv.visitVarInsn(ALOAD, 3); // workingMemory
 
                     mv.visitVarInsn(ALOAD, 5);
-                    invokeInterface(LeftTuple.class, "getFactHandle", InternalFactHandle.class);
+                    invokeVirtual(LeftTuple.class, "getFactHandle", InternalFactHandle.class);
                     invokeInterface(InternalFactHandle.class, "getObject", Object.class); // tuple.getFactHandle().getObject()
 
                     storeObjectFromDeclaration(declarations[i], expectedDeclarations[i]);
