+++ drools-mvel/src/main/java/org/drools/mvel/asm/ReturnValueGenerator.java
@@ -83,7 +83,7 @@ public class ReturnValueGenerator {
                     mv.visitVarInsn(ALOAD, 5); // workingMemory
 
                     mv.visitVarInsn(ALOAD, 7);
-                    invokeInterface(LeftTuple.class, "getFactHandle", InternalFactHandle.class);
+                    invokeVirtual(LeftTuple.class, "getFactHandle", InternalFactHandle.class);
                     invokeInterface(InternalFactHandle.class, "getObject", Object.class); // tuple.getFactHandle().getObject()
 
                     storeObjectFromDeclaration(previousDeclarations[i], previousDeclarations[i].getTypeName());
