+++ drools-ruleunit/src/main/java/org/drools/ruleunit/datasources/DataSourceFactHandle.java
@@ -26,7 +26,6 @@ import org.drools.core.common.EqualityKey;
 import org.drools.core.common.InternalFactHandle;
 import org.drools.core.common.InternalWorkingMemory;
 import org.drools.core.factmodel.traits.TraitTypeEnum;
-import org.drools.core.reteoo.BaseLeftTuple;
 import org.drools.core.reteoo.LeftTuple;
 import org.drools.core.reteoo.ObjectTypeConf;
 import org.drools.core.reteoo.RightTuple;
@@ -301,7 +300,7 @@ public class DataSourceFactHandle implements InternalFactHandle {
     }
 
     @Override
-    public void forEachLeftTuple( Consumer<BaseLeftTuple> leftTupleConsumer ) {
+    public void forEachLeftTuple( Consumer<LeftTuple> leftTupleConsumer ) {
         throw new UnsupportedOperationException( "org.drools.core.datasources.CursoredDataSource.DataSourceFactHandle.forEachLeftTuple -> TODO" );
 
     }
@@ -313,7 +312,7 @@ public class DataSourceFactHandle implements InternalFactHandle {
     }
 
     @Override
-    public BaseLeftTuple findFirstLeftTuple( Predicate<BaseLeftTuple> lefttTuplePredicate ) {
+    public LeftTuple findFirstLeftTuple(Predicate<LeftTuple> lefttTuplePredicate ) {
         throw new UnsupportedOperationException( "org.drools.core.datasources.CursoredDataSource.DataSourceFactHandle.findFirstLeftTuple -> TODO" );
 
     }
